requirements.txt
openvino==2022.2.0
openvino-dev==2022.2.0
openvino-telemetry==2022.1.2
numpy==1.19.5
Pillow==8.4.0
scipy==1.5.4
opencv-python==4.6.0.66
